MAINT: move NodalFormCapillary inside physics/capillary.py
