+++ a_package/physics/__init__.py
@@ -1,2 +1,2 @@
-from .capillary import CapillaryBridge
+from .capillary import NodalFormCapillary
 from .surfaces import generate_surface, SelfAffineRoughness, psd_to_height
