+++ a_package/simulation/__init__.py
@@ -1,6 +1,5 @@
 from .simulation import Simulation
 from .io import SimulationIO, NpyIO, Term
-from .formulation import NodalFormCapillary
 from .setup import (
     create_grid_from_config,
     generate_surface_from_config,
